+++ app/Database/Migrations/2025-08-21-121729_Substitute.php
@@ -27,8 +27,7 @@ class Substitute extends Migration
         $this->forge->createTable('substitute');
     }
 
-    public function down()
-    {
-        $this->forge->dropTable('substitute');
+    public function down() {
+        if ($this->db->tableExists('substitute')) $this->forge->dropTable('substitute');
     }
 }
