+++ app/Database/Migrations/2025-08-21-121917_Opinion.php
@@ -60,8 +60,7 @@ class Opinion extends Migration
         $this->forge->createTable('opinion');
     }
 
-    public function down()
-    {
-        $this->forge->dropTable('opinion');
+    public function down() {
+        if ($this->db->tableExists('opinion')) $this->forge->dropTable('opinion');
     }
 }
